Fix api (#3)
* Revert endpoints to 0a60afd

* Bun

* Fix Vercel deployment warnings

- Update Node.js engine to 20.x (current LTS)
- Remove deprecated memory settings from vercel.json
- Remove unnecessary maxDuration settings



* Update package scripts

* Fix package scripts

---------
