+++ api/healthcheck.ts
@@ -1,6 +1,15 @@
-import { healthcheckHandler } from './handlers/healthcheck';
+import type { VercelRequest, VercelResponse } from '@vercel/node';
 
-export default async function handler(req: Request): Promise<Response> {
-  const result = await healthcheckHandler(req.method);
-  return Response.json(result.body, { status: result.status });
+export default async function handler(req: VercelRequest, res: VercelResponse): Promise<void> {
+  // Only allow GET requests
+  if (req.method !== 'GET') {
+    res.status(405).json({ error: 'Method not allowed' });
+    return;
+  }
+
+  res.status(200).json({
+    healthcheck: Date.now(),
+    status: 'ok',
+    service: 'yearn-pricing'
+  });
 }
